+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **Changelog** Added package list in changelog output
 - **Menu** Added a package changelog selection menu to check either single, multiple or all packages
 - **Config** Added the option to change the package architecture dependent word in the config instead of in the code
+- **Config** Added the option to change the delay for the web scraping in the config instead of in the code
 
 ### Bug fixes